Implementation of authorization for view, create, update or delete a call via gates in the controller

+++ app/Http/Controllers/CallController.php
@@ -12,6 +12,7 @@ use Illuminate\Support\Facades\DB;
 
 use App\Http\Requests\CallRequest;
 use App\Http\Requests\CallSearchRequest;
+use Illuminate\Support\Facades\Gate;
 
 class CallController extends Controller
 {
@@ -29,11 +30,16 @@ class CallController extends Controller
    */
   public function index()
   {
-    $calls = Call::orderBy('id', 'desc')->paginate(5);
+    if (Gate::denies('calls.view')) {
+      return abort(403, 'Acción no autorizada.');
+    }
 
     $locale = \App::getLocale();
 
     $uri = 'call_trackings';
+
+    $calls = Call::orderBy('id', 'desc')->paginate(5);
+
 
     return view('call.index', compact('calls', 'uri'));
   }
@@ -45,6 +51,10 @@ class CallController extends Controller
    */
   public function create()
   {
+    if (Gate::denies('calls.create')) {
+      return abort(403, 'Acción no autorizada.');
+    }
+
     $locale = \App::getLocale();
 
     $uri = 'call_trackings';
@@ -65,6 +75,10 @@ class CallController extends Controller
    */
   public function store(CallRequest $request)
   {
+    if (Gate::denies('calls.create')) {
+      return abort(403, 'Acción no autorizada.');
+    }
+
     $data = $request->all();
     $data['user_id'] = \Auth::id();
     unset($data['_token']);
@@ -107,11 +121,19 @@ class CallController extends Controller
    */
   public function show(Request $request)
   {
+    if (Gate::denies('calls.view')) {
+      return abort(403, 'Acción no autorizada.');
+    }
+
     $locale = \App::getLocale();
 
     $uri = 'call_trackings';
 
     $call = Call::findOrFail($request->id);
+
+    if (Gate::denies('calls.view')) {
+      return abort(403, 'Acción no autorizada.');
+    }
 
     return view('call.show', compact('call', 'uri'));
   }
@@ -124,6 +146,10 @@ class CallController extends Controller
    */
   public function edit(Request $request)
   {
+    if (Gate::denies('calls.update')) {
+      return abort(403, 'Acción no autorizada.');
+    }
+
     $locale = \App::getLocale();
 
     $uri = 'call_trackings';
@@ -144,6 +170,10 @@ class CallController extends Controller
    */
   public function update(CallRequest $request)
   {
+    if (Gate::denies('calls.update')) {
+      return abort(403, 'Acción no autorizada.');
+    }
+
     $data = $request->all();
     unset($data['_token']);
 
@@ -186,6 +216,10 @@ class CallController extends Controller
    */
   public function destroy(Request $request)
   {
+    if (Gate::denies('calls.delete')) {
+      return abort(403, 'Acción no autorizada.');
+    }
+
     $call = Call::findOrFail($request->id);
     $destroyed = $call->delete();
 
@@ -206,6 +240,10 @@ class CallController extends Controller
 
   public function search(CallSearchRequest $request)
   {
+    if (Gate::denies('calls.view')) {
+      return abort(403, 'Acción no autorizada.');
+    }
+
     $calls = Call::whereBetween('created_at', [$request->date, now()->today()])
                  ->orderBy('id', 'desc')
                  ->paginate(5);
